feat: exception handling - delete

+++ src/main/java/br/com/sabrinaweb/project_spring_web/services/UserService.java
@@ -2,8 +2,12 @@ package br.com.sabrinaweb.project_spring_web.services;
 
 import br.com.sabrinaweb.project_spring_web.entities.User;
 import br.com.sabrinaweb.project_spring_web.repositories.UserRepository;
+import br.com.sabrinaweb.project_spring_web.services.exceptions.DataBaseException;
 import br.com.sabrinaweb.project_spring_web.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Repository;
 import org.springframework.stereotype.Service;
 
@@ -27,7 +31,16 @@ public class UserService {
         return repository.save(user);
     }
     public void delete (Long id){
-        repository.deleteById(id);
+        try {
+            if (!repository.existsById(id)) {
+                throw new ResourceNotFoundException(id);
+            }
+            else {
+                repository.deleteById(id);
+            }
+        }catch (DataIntegrityViolationException e ){
+            throw new DataBaseException(e.getMessage());
+        }
     }
     public User update(Long id, User user){
         User entity = repository.getReferenceById(id);
